requirements.txt
streamlit==1.35.0
diffusers==0.25.0
transformers==4.38.2
torch==2.1.2
accelerate==0.27.1
xformers==0.0.23.post1
ffmpeg-python==0.2.0
huggingface-hub==0.20.3